rename stack to main
